+++ Gravitas.Platform.Web/Controllers/Routine/LaboratoryInController.cs
@@ -46,7 +46,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult Idle_SelectTicketContainer(long? nodeId, long ticketContainerId)
+        public ActionResult Idle_SelectTicketContainer(int? nodeId, int ticketContainerId)
         {
             if (nodeId != null)
                 _opRoutineWebManager.LaboratoryIn_Idle_SelectTicketContainer(nodeId.Value, ticketContainerId);
@@ -55,28 +55,28 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult Idle_СollectSample(long? nodeId)
+        public ActionResult Idle_СollectSample(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_Idle_СollectSample(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
         }
 
         [HttpGet]
-        public ActionResult Idle_EditAnalysisResult(long? nodeId)
+        public ActionResult Idle_EditAnalysisResult(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_Idle_EditAnalysisResult(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
         }
 
         [HttpGet]
-        public ActionResult Idle_PrintAnalysisResult(long? nodeId)
+        public ActionResult Idle_PrintAnalysisResult(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_Idle_PrintAnalysisResult(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
         }
         
         [HttpGet]
-        public ActionResult Idle_PrintCollisionInit(long? nodeId, int ticketId)
+        public ActionResult Idle_PrintCollisionInit(int? nodeId, int ticketId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_Idle_PrintCollisionInit(nodeId.Value, ticketId);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -87,13 +87,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 02_SampleReadTruckRfid
 
         [HttpGet, ChildActionOnly]
-        public ActionResult SampleReadTruckRfid(long? nodeId) => nodeId.HasValue
+        public ActionResult SampleReadTruckRfid(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/02_SampleReadTruckRfid",
                 new LaboratoryInVms.SampleReadTruckRfidVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult SampleReadTruckRfid_Back(long? nodeId)
+        public ActionResult SampleReadTruckRfid_Back(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_SampleReadTruckRfid_Back(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -104,13 +104,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 03_SampleBindTray
 
         [HttpGet, ChildActionOnly]
-        public ActionResult SampleBindTray(long? nodeId) => nodeId.HasValue
+        public ActionResult SampleBindTray(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/03_SampleBindTray",
                 new LaboratoryInVms.SampleBindTrayVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult SampleBindTray_Back(long? nodeId)
+        public ActionResult SampleBindTray_Back(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_SampleBindTray_Back(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -121,13 +121,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 04_SampleBindAnalysisTray
 
         [HttpGet, ChildActionOnly]
-        public ActionResult SampleBindAnalysisTray(long? nodeId) => nodeId.HasValue
+        public ActionResult SampleBindAnalysisTray(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/04_SampleBindAnalysisTray",
                 _opRoutineWebManager.LaboratoryIn_SampleBindAnalysisTray_GetVmData(nodeId.Value))
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult SampleBindAnalysisTray_Next(long? nodeId)
+        public ActionResult SampleBindAnalysisTray_Next(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_SampleBindAnalysisTray_Next(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -145,7 +145,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 05_SampleAddOpVisa
 
         [HttpGet, ChildActionOnly]
-        public ActionResult SampleAddOpVisa(long? nodeId) => nodeId.HasValue
+        public ActionResult SampleAddOpVisa(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/05_SampleAddOpVisa",
                 new LaboratoryInVms.SampleAddOpVisaVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -155,13 +155,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 11_ResultReadTrayRfid
 
         [HttpGet, ChildActionOnly]
-        public ActionResult ResultReadTrayRfid(long? nodeId) => nodeId.HasValue
+        public ActionResult ResultReadTrayRfid(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/11_ResultReadTrayRfid",
                 new LaboratoryInVms.ResultReadTrayRfidVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult ResultReadTrayRfid_Back(long? nodeId)
+        public ActionResult ResultReadTrayRfid_Back(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_ResultReadTrayRfid_Back(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -172,7 +172,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 12_ResultEditAnalysis
 
         [HttpGet, ChildActionOnly]
-        public ActionResult ResultEditAnalysis(long? nodeId)
+        public ActionResult ResultEditAnalysis(int? nodeId)
         {
             if (nodeId is null) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
@@ -204,7 +204,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 13_ResultAddOpVisa
 
         [HttpGet, ChildActionOnly]
-        public ActionResult ResultAddOpVisa(long? nodeId) => nodeId.HasValue
+        public ActionResult ResultAddOpVisa(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/13_ResultAddOpVisa",
                 new LaboratoryInVms.ResultAddOpVisaVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -214,13 +214,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 21_PrintReadTrayRfid
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintReadTrayRfid(long? nodeId) => nodeId.HasValue
+        public ActionResult PrintReadTrayRfid(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/21_PrintReadTrayRfid",
                 new LaboratoryInVms.PrintReadTrayRfidVm {NodeId = nodeId.Value})
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult PrintReadTrayRfid_Back(long? nodeId)
+        public ActionResult PrintReadTrayRfid_Back(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintReadTrayRfid_Back(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -231,7 +231,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 22_PrintAnalysisResult
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintAnalysisResult(long? nodeId)
+        public ActionResult PrintAnalysisResult(int? nodeId)
         {
             if (nodeId is null) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
@@ -247,7 +247,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult PrintAnalysisResult_Back(long? nodeId)
+        public ActionResult PrintAnalysisResult_Back(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintAnalysisResult_Back(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -278,27 +278,27 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 24_PrintDataDisclose
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintDataDisclose(long? nodeId) => nodeId.HasValue
+        public ActionResult PrintDataDisclose(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/24_PrintDataDisclose",
                 _opRoutineWebManager.LaboratoryIn_PrintDataDisclose_GetVm(nodeId.Value))
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult PrintDataDisclose_Confirm(long? nodeId, bool isConfirmed)
+        public ActionResult PrintDataDisclose_Confirm(int? nodeId, bool isConfirmed)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintDataDisclose_Confirm(nodeId.Value, isConfirmed);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
         }
         
         [HttpGet]
-        public ActionResult PrintDataDisclose_Cancel(long? nodeId, string indexRefundReason)
+        public ActionResult PrintDataDisclose_Cancel(int? nodeId, string indexRefundReason)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintCollisionManage_SetReturnRoute(nodeId.Value, indexRefundReason);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
         } 
         
         [HttpGet]
-        public ActionResult PrintDataDisclose_Reload(long? nodeId)
+        public ActionResult PrintDataDisclose_Reload(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintCollisionManage_SetReloadRoute(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -317,13 +317,13 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 25_PrintCollisionInit
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintCollisionInit(long? nodeId) => nodeId.HasValue
+        public ActionResult PrintCollisionInit(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/25_PrintCollisionInit",
                 _opRoutineWebManager.LaboratoryIn_GetCollisionInitVm(nodeId.Value))
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
         [HttpGet]
-        public ActionResult PrintCollisionInit_Return(long? nodeId)
+        public ActionResult PrintCollisionInit_Return(int? nodeId)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintCollisionInit_Return(nodeId.Value);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -343,14 +343,14 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 26_PrintCollisionManage
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintCollisionManage(long? nodeId) => nodeId.HasValue
+        public ActionResult PrintCollisionManage(int? nodeId) => nodeId.HasValue
             ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/26_PrintCollisionManage",
                 _opRoutineWebManager.GetLaboratoryIn_PrintCollisionManageVm(nodeId.Value))
             : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
 
         [HttpGet]
-        public ActionResult PrintCollisionManage_ReturnToCollectSamples(long? nodeId)
+        public ActionResult PrintCollisionManage_ReturnToCollectSamples(int? nodeId)
         {
             if (nodeId != null)
                 _opRoutineWebManager.LaboratoryIn_PrintCollisionManage_ReturnToCollectSamples(nodeId.Value);
@@ -358,7 +358,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult PrintCollisionManage_SetRouteToExit(long? nodeId, string indexRefundReason)
+        public ActionResult PrintCollisionManage_SetRouteToExit(int? nodeId, string indexRefundReason)
         {
             if (nodeId != null) _opRoutineWebManager.LaboratoryIn_PrintCollisionManage_SetReturnRoute(nodeId.Value, indexRefundReason);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
@@ -379,7 +379,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 28_PrintLaboratoryProtocol
 
         [HttpGet, ChildActionOnly]
-        public ActionResult PrintLaboratoryProtocol(long? nodeId)
+        public ActionResult PrintLaboratoryProtocol(int? nodeId)
         {
             return nodeId.HasValue
                 ? (ActionResult) PartialView("../OpRoutine/LabolatoryIn/28_PrintLaboratoryProtocol",
@@ -387,7 +387,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
                 : new HttpStatusCodeResult(HttpStatusCode.BadRequest);
         }
 
-        public ActionResult PrintLaboratoryProtocol_Next(long nodeId)
+        public ActionResult PrintLaboratoryProtocol_Next(int nodeId)
         {
             _opRoutineWebManager.PrintLaboratoryProtocol_Next(nodeId);
             return new HttpStatusCodeResult(HttpStatusCode.OK);
