+++ Gravitas.Platform.Web/Controllers/Routine/LoadPointGuide2Controller.cs
@@ -21,7 +21,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         #region 01_Idle
 
         [HttpGet, ChildActionOnly]
-        public ActionResult Idle(long? nodeId)
+        public ActionResult Idle(int? nodeId)
         {
             if (nodeId == null) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
@@ -30,7 +30,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult Idle_SelectTicketContainer(long? nodeId, long ticketContainerId)
+        public ActionResult Idle_SelectTicketContainer(int? nodeId, int ticketContainerId)
         {
             if (nodeId == null) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
@@ -58,7 +58,7 @@ namespace Gravitas.Platform.Web.Controllers.Routine
         }
 
         [HttpGet]
-        public ActionResult BindLoadPoint_Back(long? nodeId)
+        public ActionResult BindLoadPoint_Back(int? nodeId)
         {
             if (nodeId == null) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
